fix and all table entity

+++ src/main/java/ru/tagallteam/hackstarter/application/achievement/model/AchievementDto.java
@@ -20,7 +20,7 @@ public class AchievementDto {
 
     private Long id;
 
-    private Integer fileId;
+    private Long fileId;
 
     private String name;
